requirements.txt
flask
plotly
dash-bio
pandas
openpyxl
numpy